Parameterise between *.xml files and Config GUI
Commit completes the overhaul of parameterisation for the first step of
the process. A configuration *.xml file can be used to store a complete
set of parameters, and loaded into the Configuration GUIs by default or
explicitly with the "Import Config" functionality. Likewise, *.xml dumps
of parameters can be created by exporting.