Support input parameters for OpenSeqSLAM2
Commit provides support for input parameters to the OpenSeqSLAM2
program. The supported parameters are:
- 'config': blank for GUI configuration, 'default' for the default
  configuration, and a string to specify a custom configuration
- 'progress': display mode for progress (either graphical or text-based)
- 'batch_param': parameter to run with a batch of different values
- 'batch_values': values to run through 'batch_param'

All parameters are supported as of this commit but not implemented.
Implementation is complete up to and including selecting the
configuration (running in text or graphical, and results for batch
operation still need to be implemented).